Add files via upload
Changed struct dn_link bandwidth to int64 for accepting values higher than 2.4Gbps

+++ sys/netinet/ip_dummynet.h
@@ -120,7 +120,9 @@ struct dn_link {
 	 * XXX what about burst ?
 	 */
 	int32_t		link_nr;
-	int		bandwidth;	/* bit/s or bits/tick.   */
+//	int		bandwidth;	/* bit/s or bits/tick.   */
+//RUDOLF - change bandwidth to 64bit, previously can hold max 2.1Gb/s only
+	int64_t		bandwidth;	/* bit/s or bits/tick.   */
 	int		delay;		/* ms and ticks */
 	uint64_t	burst;		/* scaled. bits*Hz  XXX */
 };
